Refactor blog post pages for production build
