+++ app/blog/(posts)/Birthday21/page.tsx
@@ -1,3 +1,18 @@
-export { metadata, post } from "./[id]";
-export { default } from "./[id]";
+import type { Metadata } from "next";
+import { BlogPostArticle } from "../../_components/BlogPostArticle";
+import { post } from "./post";
 
+export const metadata: Metadata = {
+  title: post.title,
+  description: post.description,
+  openGraph: {
+    title: post.title,
+    description: post.description,
+    type: "article",
+    publishedTime: post.date
+  }
+};
+
+export default function Birthday21Page() {
+  return <BlogPostArticle post={post} />;
+}
