+++ app/blog/(posts)/test/post.ts
@@ -1,0 +1,14 @@
+import type { BlogPost } from "../../posts";
+
+export const post: BlogPost = {
+  slug: "test",
+  title: "テストArtifact",
+  description: "テスト記事です",
+  date: "2025-10-21",
+  readingTime: "3 min read",
+  tags: [],
+  body: [
+    "これはテスト用の文章です",
+    "改行が正しくされていることを確認します"
+  ]
+};
